requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.12.0
    # via
    #   tensorboard
    #   tensorflow
    #   tensorflow-metadata
    #   tensorflow-transform
    #   tfx-bsl
aiohttp==3.7.4.post0
    # via
    #   gcsfs
    #   print-nanny-client
apache-beam[gcp]==2.30.0
    # via
    #   print-nanny-dataflow (setup.py)
    #   tensorflow-transform
    #   tfx-bsl
astunparse==1.6.3
    # via tensorflow
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
avro-python3==1.9.2.1
    # via apache-beam
backoff==1.10.0
    # via print-nanny-dataflow (setup.py)
cachetools==4.2.2
    # via
    #   apache-beam
    #   google-auth
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via google-crc32c
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
crcmod==1.7
    # via apache-beam
decorator==5.0.9
    # via gcsfs
dill==0.3.1.1
    # via apache-beam
docopt==0.6.2
    # via hdfs
fastavro==1.4.1
    # via apache-beam
fasteners==0.16.3
    # via google-apitools
flatbuffers==1.12
    # via tensorflow
fsspec==2021.06.0
    # via gcsfs
future==0.18.2
    # via apache-beam
gast==0.4.0
    # via tensorflow
gcsfs==2021.6.0
    # via print-nanny-dataflow (setup.py)
google-api-core[grpc,grpcgcp]==1.30.0
    # via
    #   google-api-python-client
    #   google-cloud-bigquery
    #   google-cloud-bigtable
    #   google-cloud-core
    #   google-cloud-datastore
    #   google-cloud-dlp
    #   google-cloud-language
    #   google-cloud-pubsub
    #   google-cloud-spanner
    #   google-cloud-videointelligence
    #   google-cloud-vision
google-api-python-client==1.12.8
    # via tfx-bsl
google-apitools==0.5.31
    # via apache-beam
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==0.4.4
    # via
    #   gcsfs
    #   tensorboard
google-auth==1.31.0
    # via
    #   apache-beam
    #   gcsfs
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   tensorboard
google-cloud-bigquery==2.20.0
    # via apache-beam
google-cloud-bigtable==1.7.0
    # via apache-beam
google-cloud-core==1.7.0
    # via
    #   apache-beam
    #   google-cloud-bigquery
    #   google-cloud-bigtable
    #   google-cloud-datastore
    #   google-cloud-spanner
google-cloud-datastore==1.15.3
    # via apache-beam
google-cloud-dlp==1.0.0
    # via apache-beam
google-cloud-language==1.3.0
    # via apache-beam
google-cloud-pubsub==1.7.0
    # via apache-beam
google-cloud-spanner==1.19.1
    # via apache-beam
google-cloud-videointelligence==1.16.1
    # via apache-beam
google-cloud-vision==1.0.0
    # via apache-beam
google-crc32c==1.1.2
    # via google-resumable-media
google-pasta==0.2.0
    # via tensorflow
google-resumable-media==1.3.0
    # via google-cloud-bigquery
googleapis-common-protos[grpc]==1.53.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   tensorflow-metadata
grpc-google-iam-v1==0.12.3
    # via
    #   google-cloud-bigtable
    #   google-cloud-pubsub
    #   google-cloud-spanner
grpcio-gcp==0.2.2
    # via
    #   apache-beam
    #   google-api-core
grpcio==1.34.1
    # via
    #   apache-beam
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-gcp
    #   tensorboard
    #   tensorflow
    #   tensorflow-serving-api
h5py==3.1.0
    # via tensorflow
hdfs==2.6.0
    # via apache-beam
httplib2==0.19.1
    # via
    #   apache-beam
    #   google-api-python-client
    #   google-apitools
    #   google-auth-httplib2
    #   oauth2client
idna==2.10
    # via
    #   requests
    #   yarl
keras-nightly==2.5.0.dev2021032900
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
markdown==3.3.4
    # via tensorboard
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
nptyping==1.4.2
    # via print-nanny-dataflow (setup.py)
numpy==1.19.5
    # via
    #   apache-beam
    #   h5py
    #   keras-preprocessing
    #   nptyping
    #   opt-einsum
    #   pandas
    #   print-nanny-dataflow (setup.py)
    #   pyarrow
    #   tensorboard
    #   tensorflow
    #   tensorflow-transform
    #   tfx-bsl
oauth2client==4.1.3
    # via
    #   apache-beam
    #   google-apitools
oauthlib==3.1.1
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==20.9
    # via
    #   google-api-core
    #   google-cloud-bigquery
pandas==1.2.4
    # via
    #   print-nanny-dataflow (setup.py)
    #   tfx-bsl
pillow==8.2.0
    # via print-nanny-dataflow (setup.py)
print-nanny-client==0.7.0.dev24
    # via print-nanny-dataflow (setup.py)
proto-plus==1.18.1
    # via google-cloud-bigquery
protobuf==3.17.3
    # via
    #   apache-beam
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   print-nanny-client
    #   proto-plus
    #   tensorboard
    #   tensorflow
    #   tensorflow-metadata
    #   tensorflow-serving-api
    #   tensorflow-transform
    #   tfx-bsl
pyarrow==2.0.0
    # via
    #   apache-beam
    #   print-nanny-dataflow (setup.py)
    #   tensorflow-transform
    #   tfx-bsl
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via cffi
pydot==1.4.2
    # via
    #   apache-beam
    #   tensorflow-transform
pymongo==3.11.4
    # via apache-beam
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
    #   pydot
python-dateutil==2.8.1
    # via
    #   apache-beam
    #   pandas
    #   print-nanny-client
pytz==2021.1
    # via
    #   apache-beam
    #   google-api-core
    #   pandas
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   apache-beam
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   hdfs
    #   requests-oauthlib
    #   tensorboard
rsa==4.7.2
    # via
    #   google-auth
    #   oauth2client
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   fasteners
    #   google-api-core
    #   google-api-python-client
    #   google-apitools
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-pasta
    #   google-resumable-media
    #   grpcio
    #   hdfs
    #   keras-preprocessing
    #   oauth2client
    #   print-nanny-client
    #   protobuf
    #   python-dateutil
    #   tensorflow
    #   tensorflow-transform
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard==2.5.0
    # via tensorflow
tensorflow-estimator==2.5.0
    # via tensorflow
tensorflow-metadata==1.0.0
    # via
    #   tensorflow-transform
    #   tfx-bsl
tensorflow-serving-api==2.5.1
    # via tfx-bsl
tensorflow-transform==1.0.0
    # via print-nanny-dataflow (setup.py)
tensorflow==2.5.0
    # via
    #   print-nanny-dataflow (setup.py)
    #   tensorflow-serving-api
    #   tensorflow-transform
    #   tfx-bsl
termcolor==1.1.0
    # via tensorflow
tfx-bsl==1.0.0
    # via tensorflow-transform
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   apache-beam
    #   tensorflow
typish==1.9.2
    # via nptyping
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.5
    # via
    #   print-nanny-client
    #   requests
werkzeug==2.0.1
    # via tensorboard
wheel==0.36.2
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.12.1
    # via tensorflow
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
